Fixed a minor bug in the format of the amount_quote variable

+++ scripts/ashbringer_config.py
@@ -39,5 +39,5 @@ class ExcaliburConfig(StrategyV2ConfigBase):
     # Triple Barrier
 
     # Order settings
-    amount_quote: int = 30
+    amount_quote: Decimal = 30.0
     min_price_delta_pct_to_open_tr: Decimal = 5.0
